pyproject.toml
[tool.poetry]
name = "newfastapi"
version = "0.1.0"
description = ""
authors = ["YaroslavPich <pich.yaroslav@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
pydantic = "^2.7.4"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.30.1"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
fastapi-jwt-auth = "^0.5.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.1.3"
aioredis = "2.0.0"
celery = "^5.4.0"
python-dotenv = "^1.0.1"
redis = "^5.0.7"
setuptools = "^70.2.0"
asyncio-redis = "^0.16.0"
fastapi-mail = "^1.4.1"
asyncio = "^3.4.3"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.40.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
